Improvements on materials
