+++ .clang-tidy
@@ -8,6 +8,7 @@ Checks: clang-diagnostic-*
     hicpp-*
     -cppcoreguidelines-pro-bounds-constant-array-index
     -modernize-use-trailing-return-type
+    -cppcoreguidelines-avoid-non-const-global-variables
     -hicpp-signed-bitwise
     -hicpp-vararg
     -hicpp-braces-around-statements
