+++ rawrbox.code-workspace
@@ -18,7 +18,6 @@
                 "query": "is:open"
             }
         ],
-        "clangd.serverCompletionRanking": false,
         "clangd.restartAfterCrash": true,
         "editor.bracketPairColorization.enabled": true,
         "editor.formatOnSave": true,
