+++ render/include/rawrbox/render/model/material/skinned_unlit.hpp
@@ -1,8 +1,6 @@
 #pragma once
 
-#include <rawrbox/render/model/light/manager.hpp>
 #include <rawrbox/render/model/material/base.hpp>
-#include <rawrbox/render/model/material/unlit.hpp>
 #include <rawrbox/render/shader_defines.hpp>
 #include <rawrbox/render/static.hpp>
 
@@ -10,8 +8,6 @@
 
 #include <bgfx/bgfx.h>
 #include <fmt/format.h>
-
-#include <vector>
 
 // NOLINTBEGIN(*)
 static const bgfx::EmbeddedShader model_skinned_unlit_shaders[] = {
@@ -22,13 +18,32 @@ static const bgfx::EmbeddedShader model_skinned_unlit_shaders[] = {
 
 namespace rawrBox {
 
-	class MaterialSkinnedUnlit : public rawrBox::MaterialUnlit {
-
+	class MaterialSkinnedUnlit : public rawrBox::MaterialBase {
 	public:
-		using MaterialUnlit::MaterialUnlit;
+		bgfx::UniformHandle u_bones = BGFX_INVALID_HANDLE;
 
-		void upload() override {
-			this->buildShader(model_skinned_unlit_shaders, "model_skinned_unlit");
+		using vertexBufferType = rawrBox::VertexSkinnedUnlitData;
+
+		MaterialSkinnedUnlit() = default;
+		MaterialSkinnedUnlit(MaterialSkinnedUnlit&&) = delete;
+		MaterialSkinnedUnlit& operator=(MaterialSkinnedUnlit&&) = delete;
+		MaterialSkinnedUnlit(const MaterialSkinnedUnlit&) = delete;
+		MaterialSkinnedUnlit& operator=(const MaterialSkinnedUnlit&) = delete;
+		~MaterialSkinnedUnlit() {
+			RAWRBOX_DESTROY(u_bones);
+		}
+
+		void registerUniforms() {
+			MaterialBase::registerUniforms();
+
+			// BONES ----
+			u_bones = bgfx::createUniform("u_bones", bgfx::UniformType::Mat4, rawrBox::MAX_BONES_PER_MODEL);
+			// ---
+		}
+
+		void upload() {
+			buildShader(model_skinned_unlit_shaders, "model_skinned_unlit");
 		}
 	};
+
 } // namespace rawrBox
