+++ render/include/rawrbox/render/static.hpp
@@ -14,7 +14,6 @@
 	}
 // NOLINTEND(*)
 namespace rawrBox {
-	// NOLINTBEGIN{cppcoreguidelines-avoid-non-const-global-variables}
 	constexpr auto MAX_BONES_PER_VERTEX = 4;
 	constexpr auto MAX_BONES_PER_MODEL = 200;
 	constexpr auto MAX_LIGHTS = 12;
@@ -23,5 +22,4 @@ namespace rawrBox {
 	extern std::shared_ptr<rawrBox::TextureMissing> MISSING_TEXTURE;
 	extern std::shared_ptr<rawrBox::TextureFlat> MISSING_SPECULAR_TEXTURE;
 	extern std::shared_ptr<rawrBox::TextureFlat> WHITE_TEXTURE;
-	// NOLINTEND{cppcoreguidelines-avoid-non-const-global-variables}
 } // namespace rawrBox
