+++ render/shaders/model/lit/vs_model_lit.sc
@@ -1,14 +1,13 @@
-$input a_position, a_normal, a_tangent, a_color0, a_texcoord0, a_indices, a_weight
+$input a_position, a_normal, a_tangent, a_color0, a_texcoord0
 $output v_color0, v_texcoord0, v_wPos, v_normal, v_tangent, v_bitangent
 
 #include <bgfx_shader.sh>
-#include <../include/model_psx.sh>
-#include <../include/model_skinned.sh>
+#include <../../include/model_psx.sh>
 
 void main() {
-	vec4 world = mul(u_model[0], boneTransform(a_indices, a_weight, a_position) );
+	vec4 world = mul(u_model[0], vec4(a_position, 1.0) );
 
-	vec4 translated = mul(u_viewProj, world);
+    vec4 translated = mul(u_viewProj, world);
 	gl_Position = psx_snap(translated, u_viewRect.zw / 2.);
 
     v_color0 = a_color0;
