+++ render/shaders/model/skinned/lit/varying.def.sc
@@ -1,8 +1,8 @@
 vec3 a_position         : POSITION;
-vec4 a_normal           : NORMAL;
-vec4 a_tangent          : TANGENT;
 vec2 a_texcoord0        : TEXCOORD0;
 vec4 a_color0           : COLOR0;
+vec4 a_normal           : NORMAL;
+vec4 a_tangent          : TANGENT;
 ivec4 a_indices         : BLENDINDICES0;
 vec4 a_weight           : BLENDWEIGHT0;
