+++ render/shaders/model/skinned/unlit/fs_model_skinned_unlit.sc
