+++ render/shaders/model/skinned/unlit/varying.def.sc
@@ -1,0 +1,9 @@
+vec3 a_position         : POSITION;
+vec2 a_texcoord0        : TEXCOORD0;
+vec4 a_color0           : COLOR0;
+ivec4 a_indices         : BLENDINDICES0;
+vec4 a_weight           : BLENDWEIGHT0;
+
+vec2  v_texcoord0       : TEXCOORD0     = vec2(0.0, 0.0);
+vec4  v_color0    	    : COLOR0        = vec4(1.0, 0.0, 0.0, 0.0);
+
