+++ render/shaders/model/unlit/vs_model_unlit.sc
@@ -2,7 +2,7 @@ $input a_position, a_color0, a_texcoord0
 $output v_color0, v_texcoord0
 
 #include <bgfx_shader.sh>
-#include <../include/model_psx.sh>
+#include <../../include/model_psx.sh>
 
 void main() {
     vec4 translated = mul(u_modelViewProj, vec4(a_position, 1.0));
