+++ render/shaders/model/unlit/vs_sprite_unlit.sc
@@ -2,7 +2,7 @@ $input a_position, a_color0, a_texcoord0
 $output v_color0, v_texcoord0
 
 #include <bgfx_shader.sh>
-#include <../include/model_psx.sh>
+#include <../../include/model_psx.sh>
 
 uniform vec3 u_sprite_pos;
