+++ samples/002-generated-models/src/game.cpp
@@ -1,6 +1,4 @@
 
-#include <rawrbox/render/model/material/lit.hpp>
-#include <rawrbox/render/model/material/unlit.hpp>
 #include <rawrbox/render/model/mesh.hpp>
 #include <rawrbox/utils/keys.hpp>
 
@@ -61,7 +59,7 @@ namespace model {
 		// Setup camera
 		this->_camera = std::make_shared<rawrBox::CameraPerspective>(static_cast<float>(width) / static_cast<float>(height), 60.0F, 0.1F, 100.0F, bgfx::getCaps()->homogeneousDepth);
 		this->_camera->setPos({0.F, 5.F, -5.F});
-		this->_camera->setAngle({0.F, bx::toRad(-45), 0.F, 0.F});
+		this->_camera->setAngle({0.F, 0.F, bx::toRad(-45), 0.F});
 		// --------------
 
 		// Load content ---
@@ -79,39 +77,44 @@ namespace model {
 		this->_texture2 = std::make_shared<rawrBox::TextureGIF>("./content/textures/meow3.gif");
 		this->_texture2->upload();
 
-		auto mat = std::make_shared<rawrBox::MaterialUnlit>();
-		this->_model = std::make_shared<rawrBox::Model>(mat);
 		// ----
-
 		{
-			auto mesh = rawrBox::ModelBase::generatePlane({5, 0, 0}, {0.5F, 0.5F});
+			auto mesh = this->_model->generatePlane({5, 0, 0}, {0.5F, 0.5F});
 			mesh->setTexture(this->_texture);
 			this->_model->addMesh(mesh);
 		}
 
 		{
-			auto mesh = rawrBox::ModelBase::generateCube({-5, 0, 0}, {0.5F, 0.5F}, rawrBox::Colors::White);
+			auto mesh = this->_model->generateCube({-5, 0, 0}, {0.5F, 0.5F}, rawrBox::Colors::White);
 			mesh->setTexture(this->_texture2);
 			this->_model->addMesh(mesh);
 		}
 
 		{
-			auto mesh = rawrBox::ModelBase::generateAxis(1, {0.F, 0.F, 0.F});
+			auto mesh = this->_model->generateAxis(1, {0.F, 0.F, 0.F});
 			this->_model->addMesh(mesh);
 		}
 
 		{
-			auto mesh = rawrBox::ModelBase::generateGrid(12, {0.F, -2.0F, 0.F});
+			auto mesh = this->_model->generateGrid(12, {0.F, 0.F, 0.F});
 			this->_model->addMesh(mesh);
 		}
 
+		{
+			auto mesh = this->_sprite->generatePlane({0, 2, 0}, {0.2F, 0.2F});
+			mesh->setTexture(this->_texture);
+			this->_sprite->addMesh(mesh);
+		}
+
 		this->_model->upload();
+		this->_sprite->upload();
 		// -----
 	}
 
 	void Game::shutdown() {
 		this->_render = nullptr;
 		this->_model = nullptr;
+		this->_sprite = nullptr;
 
 		rawrBox::Engine::shutdown();
 	}
@@ -165,6 +168,7 @@ namespace model {
 		if (this->_model == nullptr) return;
 
 		this->_model->draw(this->_camera->getPos());
+		this->_sprite->draw(this->_camera->getPos());
 	}
 
 	void Game::draw(const double alpha) {
